requirements.txt
elasticsearch
elasticsearch-dsl
sentence-transformers
flask
numpy
zmq
attrs